# src/generated/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HarborGuard API",
    "version": "0.1b",
    "description": "HarborGuard Container Security Platform API Documentation",
    "contact": {
      "name": "HarborGuard Team",
      "url": "https://github.com/HarborGuard/HarborGuard"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/HarborGuard/HarborGuard/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/api/images/export": {
      "post": {
        "summary": "POST /api/images/export",
        "tags": [
          "Images"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tarPath": {
                    "type": "string",
                    "description": "tarPath parameter"
                  },
                  "imageName": {
                    "type": "string",
                    "description": "imageName parameter"
                  },
                  "imageTag": {
                    "type": "string",
                    "description": "imageTag parameter"
                  },
                  "sourceImage": {
                    "type": "string",
                    "description": "sourceImage parameter"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/images/export-to-registry": {
      "post": {
        "summary": "POST /api/images/export-to-registry",
        "tags": [
          "Images"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sourceImage": {
                    "type": "string",
                    "description": "sourceImage parameter"
                  },
                  "targetRegistry": {
                    "type": "string",
                    "description": "targetRegistry parameter"
                  },
                  "targetImageName": {
                    "type": "string",
                    "description": "targetImageName parameter"
                  },
                  "targetImageTag = 'latest'": {
                    "type": "string",
                    "description": "targetImageTag = 'latest' parameter"
                  },
                  "repositoryId": {
                    "type": "string",
                    "description": "repositoryId parameter"
                  },
                  "patchedTarPath": {
                    "type": "string",
                    "description": "patchedTarPath parameter"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/images/name/{name}": {
      "get": {
        "summary": "GET /api/images/name/{name}",
        "tags": [
          "Images"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "name parameter"
          }
        ]
      },
      "delete": {
        "summary": "DELETE /api/images/name/{name}",
        "tags": [
          "Images"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "name parameter"
          }
        ]
      }
    },
...
    "/api/repositories/{id}/images": {
      "get": {
        "summary": "GET /api/repositories/{id}/images",
        "tags": [
          "Repositories"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id parameter"
          }
        ]
      }
    },
...
  },
...
}